Remove TODO related to noindex checks
This should be covered in a separate module as the functionality is a
bit different. This will also make the reporting simpler vs handling
additional combinations of results from both the robots.txt check and
the page checks. Can create a helper function to run both types of
checks (and any others) at the same time in future.